Moved code from header into cpp file
